+++ FifthHomework12.10.21/PostfixCalculator/CMakeLists.txt
@@ -6,4 +6,4 @@ set(CMAKE_C_STANDARD 99)
 include_directories(../Stack)
 
 add_executable(PostfixCalculator main.c ../Stack/stackTests.c
-        ../Stack/stack.c ../Stack/stackTests.h ../Stack/stack.h)
+        ../Stack/stack.c ../Stack/stackTests.h ../Stack/stack.h postfixCalculator.h postfixCalculatorTests.h postfixCalculator.c postfixCalculatorTests.c)
