+++ FifthHomework12.10.21/PostfixCalculator/postfixCalculatorTests.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <stdbool.h>
+
+// tests the postfix calculator
+bool testPostfixCalculator();
